requirements.txt
pika==1.1.0
python-dotenv==0.13.0